# Conflicts:
#	MVC_Project.WebBackend/Controllers/InvoicingController.cs
#	MVC_Project.WebBackend/Scripts/views/Invoicing/Invoice.js
#	MVC_Project.WebBackend/Views/Invoicing/Invoice.cshtml
#	MVC_Project.WebBackend/Web.config

+++ MVC_Project.Domain/Entities/RateFee.cs
@@ -10,7 +10,7 @@ namespace MVC_Project.Domain.Entities
     {
         public virtual Int64 id { get; set; }
         public virtual string type { get; set; }
-        public virtual decimal maximumValue { get; set; }
+        public virtual double maximumValue { get; set; }
         public virtual string taxes { get; set; }
         public virtual string factor { get; set; }
         public virtual bool transfer { get; set; }
